Remove access control v1 classes.
