+++ service/src/main/java/bio/terra/tanagra/service/accesscontrol/impl/OpenAccessControl.java
@@ -1,31 +1,25 @@
 package bio.terra.tanagra.service.accesscontrol.impl;
 
-import bio.terra.tanagra.service.accesscontrol.AccessControl;
 import bio.terra.tanagra.service.accesscontrol.Permissions;
 import bio.terra.tanagra.service.accesscontrol.ResourceCollection;
 import bio.terra.tanagra.service.accesscontrol.ResourceId;
 import bio.terra.tanagra.service.accesscontrol.ResourceType;
+import bio.terra.tanagra.service.accesscontrol.UnderlayAccessControl;
 import bio.terra.tanagra.service.authentication.UserId;
-import javax.annotation.Nullable;
 
-/**
- * Open access control plugin implementation that allows everything: all actions, listing all
- * resources.
- */
-public class OpenAccessControl implements AccessControl {
+public class OpenAccessControl implements UnderlayAccessControl {
   @Override
   public String getDescription() {
-    return "Allows access to any resource by anyone";
+    return "Open access control -- no authorization checks";
   }
 
   @Override
-  public boolean isAuthorized(UserId user, Permissions permissions, @Nullable ResourceId resource) {
-    return true;
+  public ResourceCollection listUnderlays(UserId user, int offset, int limit) {
+    return ResourceCollection.allResourcesAllPermissions(ResourceType.UNDERLAY, null);
   }
 
   @Override
-  public ResourceCollection listAllPermissions(
-      UserId user, ResourceType type, @Nullable ResourceId parentResource, int offset, int limit) {
-    return ResourceCollection.allResourcesAllPermissions(type, parentResource);
+  public Permissions getUnderlay(UserId user, ResourceId underlay) {
+    return Permissions.allActions(ResourceType.UNDERLAY);
   }
 }
