+++ service/src/test/java/bio/terra/tanagra/service/accesscontrol/impl/MockVerilyGroupsAccessControl.java
@@ -1,6 +1,5 @@
 package bio.terra.tanagra.service.accesscontrol.impl;
 
-import bio.terra.tanagra.service.accesscontrol2.impl.VerilyGroupsAccessControl;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
